Split apart crafting / mod.rs
