+++ emergence_lib/src/player_interaction/clipboard.rs
@@ -6,7 +6,7 @@ use leafwing_input_manager::prelude::ActionState;
 
 use crate::{
     asset_management::manifest::Id,
-    crafting::ActiveRecipe,
+    crafting::components::ActiveRecipe,
     simulation::geometry::{Facing, MapGeometry, TilePos},
     structures::{construction::Preview, structure_manifest::Structure},
 };
