+++ emergence_lib/src/structures/construction.rs
@@ -4,7 +4,7 @@
 //! Ghosts are buildings that are genuinely planned to be built.
 //! Previews are simply hovered, and used as a visual aid to show placement.
 
-use crate::crafting::WorkersPresent;
+use crate::crafting::components::WorkersPresent;
 use crate::simulation::geometry::MapGeometry;
 use crate::terrain::terrain_manifest::Terrain;
 use crate::{self as emergence_lib, graphics::InheritedMaterial};
@@ -18,7 +18,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     asset_management::manifest::Id,
-    crafting::{ActiveRecipe, CraftingState, InputInventory},
+    crafting::components::{ActiveRecipe, CraftingState, InputInventory},
     player_interaction::clipboard::ClipboardData,
     signals::{Emitter, SignalStrength, SignalType},
     simulation::geometry::{Facing, TilePos},
